+++ src/rcer_iot_client_pkg/general_types/error_types/common/__init__.py
@@ -1,7 +1,7 @@
-from src.rcer_iot_client_pkg.general_types.error_types.common.common_types import (
+from .common_types import (
     EmptyDataError,
     FtpClientError,
-    HttpClientError,
+    SharepointClientError,
 )
 
-__all__ = ["EmptyDataError", "HttpClientError", "FtpClientError"]
+__all__ = ["EmptyDataError", "SharepointClientError", "FtpClientError"]