+++ src/rcer_iot_client_pkg/libs/ftp_client/__init__.py
@@ -1,4 +1,4 @@
 from .ftp_client import FTPClient
-from .types import FtpClientInitArgs, ListFilesArgs, ReadFileArgs
+from .types import FtpClientInitArgs, FtpListFilesArgs, FtpReadFileArgs
 
-__all__ = ["FTPClient", "FtpClientInitArgs", "ListFilesArgs", "ReadFileArgs"]
+__all__ = ["FTPClient", "FtpClientInitArgs", "FtpListFilesArgs", "FtpReadFileArgs"]