+++ src/rcer_iot_client_pkg/libs/ftp_client/types/ftp_client_types.py
@@ -11,10 +11,10 @@ class FtpClientInitArgs:
 
 
 @dataclass
-class ListFilesArgs:
+class FtpListFilesArgs:
     path: str
 
 
 @dataclass
-class ReadFileArgs:
+class FtpReadFileArgs:
     file_path: str