+++ src/rcer_iot_client_pkg/libs/sharepoint_client/sharepoint_client_contract.py
@@ -1,16 +1,22 @@
-
-from .types.sharepoint_client_types import ListFilesArgs, UploadFileArgs
 from abc import ABC, abstractmethod
+
+from .types.sharepoint_client_types import (
+    SpListFilesArgs,
+    SpListFoldersArgs,
+    SpUploadFileArgs,
+)
+
 
 class SharepointClientContract(ABC):
     @abstractmethod
-    async def list_files(self, args: ListFilesArgs) -> list:
+    async def list_files(self, args: SpListFilesArgs) -> list:
         pass
-    
+
     @abstractmethod
-    async def list_folders(self, args: ListFilesArgs) -> list: 
+    async def list_folders(self, args: SpListFoldersArgs) -> list:
         pass
-    
+
     @abstractmethod
-    async def upload_file(self, args: UploadFileArgs) -> dict: 
+    async def upload_file(self, args: SpUploadFileArgs) -> dict:
         pass
+