+++ src/rcer_iot_client_pkg/libs/sharepoint_client/types/sharepoint_client_types.py
@@ -1,19 +1,23 @@
 from dataclasses import dataclass
 
+
 @dataclass
 class SharepointClientInitArgs:
     client_name: str = "sharepoint_rest_api"
-    
+
+
 @dataclass
-class ListFilesArgs: 
+class SpListFilesArgs:
     folder_relative_url: str
-    
+
+
 @dataclass
-class ListFoldersArgs:
+class SpListFoldersArgs:
     folder_relative_url: str
-    
+
+
 @dataclass
-class UploadFileArgs:
+class SpUploadFileArgs:
     file_path: str
     folder_relative_url: str
     file_content: bytes = bytes()