+++ src/rcer_iot_client_pkg/services/epii/api.py
@@ -1,10 +1,11 @@
 from typing import Dict
 
-from src.rcer_iot_client_pkg.services.epii.controllers import UpdateThiesDataController
-from src.rcer_iot_client_pkg.services.epii.controllers.types import UpdateThiesDataControllerInput
+from .controllers.types.update_thies_data_types import UpdateThiesDataControllerInput
+from .controllers.update_thies_data import UpdateThiesDataController
+
 
 class EpiiAPI:
-    def update_thies_data(
+    async def update_thies_data(
         self,
         ftp_port: int,
         ftp_host: str,
@@ -19,5 +20,5 @@ class EpiiAPI:
                 ftp_user=ftp_user,
             )
         )
-        response = controller.execute()
+        response = await controller.execute()
         return response.__dict__