+++ src/rcer_iot_client_pkg/services/epii/controllers/__init__.py
@@ -1,5 +1,3 @@
-from src.rcer_iot_client_pkg.services.epii.controllers.update_thies_data import (
-    UpdateThiesDataController,
-)
+from .update_thies_data import UpdateThiesDataController
 
 __all__ = ["UpdateThiesDataController"]