+++ tests/epii/test_api.py
@@ -1,16 +1,16 @@
 import unittest
-from unittest.mock import Mock, patch
+from unittest.mock import AsyncMock, patch
 
-from src.rcer_iot_client_pkg.services.epii.api import EpiiAPI
-from src.rcer_iot_client_pkg.services.epii.controllers.types import (
+from rcer_iot_client_pkg.services.epii.api import EpiiAPI
+from rcer_iot_client_pkg.services.epii.controllers.types import (
     UpdateThiesDataControllerOutput,
 )
 
 
-class TestEpiiAPIUpdateThiesData(unittest.TestCase):
-    @patch("src.rcer_iot_client_pkg.services.epii.api.UpdateThiesDataController")
-    def test_should_update_thies_data_successfully(
-        self, mock_update_thies_data_controller: Mock
+class TestEpiiAPIUpdateThiesData(unittest.IsolatedAsyncioTestCase):
+    @patch("rcer_iot_client_pkg.services.epii.api.UpdateThiesDataController")
+    async def test_should_update_thies_data_successfully(
+        self, mock_update_thies_data_controller
     ):
         # Arrange
         api = EpiiAPI()
@@ -21,17 +21,20 @@ class TestEpiiAPIUpdateThiesData(unittest.TestCase):
         expected_response = UpdateThiesDataControllerOutput(
             message="valid message", status=200, metadata={"data": "value"}
         )
-        mock_update_thies_data_controller_inst = (
-            mock_update_thies_data_controller.return_value
+
+        mock_update_thies_data_controller_inst = mock_update_thies_data_controller.return_value
+        mock_update_thies_data_controller_inst.execute = AsyncMock(
+            return_value=expected_response
         )
-        mock_update_thies_data_controller_inst.execute.return_value = expected_response
+
         # Act
-        response = api.update_thies_data(
+        response = await api.update_thies_data(
             ftp_port=ftp_port,
             ftp_host=ftp_host,
             ftp_password=ftp_password,
             ftp_user=ftp_user,
         )
+
         # Assert
         self.assertEqual(response, expected_response.__dict__)
         mock_update_thies_data_controller_inst.execute.assert_called_once()