+++ CHANGELOG.md
@@ -1,7 +1,52 @@
-# Changelog
+# CHANGELOG
 
-<!--next-version-placeholder-->
 
-## v0.1.0 (16/04/2025)
+## v0.3.0 (2025-04-17)
 
-- First release of `rcer_iot_client_pkg`!
+
+## v0.2.0 (2025-04-17)
+
+
+## v0.1.0 (2025-04-17)
+
+### Bug Fixes
+
+- Absolute paths
+  ([`a7a849a`](https://github.com/pedrozavalat/rcer_iot_client_pkg/commit/a7a849a2b82b32a3b25af0a4c2f9feb97b5f2994))
+
+- Add ruff step
+  ([`39cec99`](https://github.com/pedrozavalat/rcer_iot_client_pkg/commit/39cec9971e45e45edcb1dc93306c1d9d505ed597))
+
+- Bug environment file
+  ([`b15ed80`](https://github.com/pedrozavalat/rcer_iot_client_pkg/commit/b15ed80a9d473deac8a6d7fd5dc03bab3a5cef4b))
+
+- Ci/cd
+  ([`a8191ff`](https://github.com/pedrozavalat/rcer_iot_client_pkg/commit/a8191ff4a0c7c4c111139a444352a62ad4cb0fd3))
+
+- Ci/cd
+  ([`bfee434`](https://github.com/pedrozavalat/rcer_iot_client_pkg/commit/bfee434502f6a3d1b460343fecfbc7bf33618e0e))
+
+- Ci/cd
+  ([`d2530bd`](https://github.com/pedrozavalat/rcer_iot_client_pkg/commit/d2530bde0d38bff6ce30ef1d76b3d878e61fd118))
+
+- Remove documentation step
+  ([`45c1bba`](https://github.com/pedrozavalat/rcer_iot_client_pkg/commit/45c1bba37a5078a9660290e14921417062fe9f55))
+
+- **tests**: Debug absolute paths
+  ([`07eb3a5`](https://github.com/pedrozavalat/rcer_iot_client_pkg/commit/07eb3a5781c208673224a1465512c8a2c51098bf))
+
+### Features
+
+- Realease 0.1.2
+  ([`3335019`](https://github.com/pedrozavalat/rcer_iot_client_pkg/commit/3335019524d68b7b4a39c7b0983b4d7ca1543f6d))
+
+- Update pyproject
+  ([`d00225b`](https://github.com/pedrozavalat/rcer_iot_client_pkg/commit/d00225be0ba67dd2f2d607ab36a5f47cf57d9927))
+
+- Upload initial files
+  ([`df27dfc`](https://github.com/pedrozavalat/rcer_iot_client_pkg/commit/df27dfc826a031602585b3c5017907b13c64a48a))
+
+### Refactoring
+
+- **lint**: Clean code
+  ([`512cebc`](https://github.com/pedrozavalat/rcer_iot_client_pkg/commit/512cebc86f0252d5a973a5c5b20db03221a0f7c9))